[WICS -6] pagenumber param changed to requestparam from pathvariable
