+++ wic/src/main/webapp/WEB-INF/tiles/skeleton.jsp
@@ -59,10 +59,15 @@
           	 <li class="dropdown">
 	          <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">Import <span class="caret"></span></a>
 	          <ul class="dropdown-menu">
-	            <li><a href="<c:url value="/admin/text/person" />">Import Person</a></li>
-	            <li><a href="<c:url value="/admin/text/list" />">Show Imports</a></li>
+	            <li><a href="<c:url value="/admin/import/person" />">Import Person</a></li>
+	            <li><a href="<c:url value="/admin/import/list" />">Show Imports</a></li>
+	          </ul>
+	        </li>
+	        <li class="dropdown">
+	          <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">Text <span class="caret"></span></a>
+	          <ul class="dropdown-menu">
 	            <li><a href="<c:url value="/admin/text/addTextView" />">Add Text</a></li>
-	            <li><a href="<c:url value="/admin/text/list/1" />">Show Text</a></li>
+	            <li><a href="<c:url value="/admin/text/list?pageNumber=1" />">Show Text</a></li>
 	          </ul>
 	        </li>
             </sec:authorize>
